requirements.txt
openai==0.27.0
tiktoken==0.3.0
structlog==22.3.0
cachetools==5.2.0
IPython
gtts
PyGithub
retrying
bs4
feedparser
PyPDF2
Pillow
huggingface_hub
geopy
numpy
numba
dalle3
timezonefinder
pytz
